+++ src/Cart/Arrayable.php
@@ -2,7 +2,7 @@
 
 namespace Cart;
 
-interface ArrayableInterface
+interface Arrayable
 {
     /**
      * Get the instance as an array.