+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 #Changelog
 
+##2.1.1
+
+- renamed `Cart\ArrayableInterface` to `Cart\Arrayable`
+- added scrutinizer config
+- fixed scrutinizer recommendations
+
 ##2.1.0
 
 - Renamed `Cart\Store\Native\SessionStore` to `Cart\SessionStore`